Generate a human(ish) description of report period
If the reporting period was in the last seven days, we generally say
"Since Monday [morning|afternoon]" or "On Monday". The former is used
when the end of the reporting period was less than 24 hours ago. The
latter is used when the reporting period spans a single day.

For older reports (or complicated ranges) we use  longer descriptions:
"On Wednesday 1 January" or
"From Wednesday 1 January to Friday 3 January".

Closes #16.